memtx: do not set NULL tuple in tree iterator
Currently, we set NULL tuple to iterator when it is exhausted. But we will
need this tuple for pagination, so let's leave last tuple in iterator.

This patch also renames iterator's field currrent to last to denote a new
property.

NO_TEST=no behaviour changes
NO_CHANGELOG=internal
NO_DOC=internal

Part of #7633