tuple_extract: allocate enough memory in tuple_extract_key_slowpath_raw
Currently, tuple_extract_key_raw is used only for primary key,
which is not nullable, so the problem was not observed before.
Its slowpath version chunk of the same size as tuple, but if
the tuple has trailing NULL fields, allocated chunk is not enough.
The patch fixes this problem.
Also, there is a special assert in this function which is poorly
located - let's move it to the end of the funciton.

Part of #7633

NO_CHANGELOG=internal
NO_DOC=internal