replication: invalid recovery from xlog
If you shut down the replica and delete all xlog files on it, and it
reconnect to the master, may occur the restore error. Because there are
intermediate xlog files between the snap on the replica and the last
xlog on the master. When restoring xlog files on the replica, it can be
occurred that prev_signature < signature and this will lead to the crash
of the master.

Closes #5158

NO_DOC=bugfix

(cherry picked from commit f53fde1)