test: factor wal play until "synchro busy" out
The function play_wal_until_synchro_queue_is_busy() was used in a
few tests copy-pasted since it was considered to be too specific
for a few rare tests. But apparently it is going to be used again
in a new test in a future commit.

The patch makes this function a method of server object to reuse
it properly.

Needed for #7253

NO_DOC=refactoring
NO_CHANGELOG=refactoring