index: make index_read_view_iterator fixed size
This is a straightforward patch that makes the read view iterator struct
fixed-size so that it can be allocated on stack. This is a pre-requisite
for the raw C API for read views: the function that creates an iterator
should allocate no memory.

Closes #7813

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring