debugger: swap cmd_up and cmd_down commands
Historically luadebug.lua used `up` and `down` in
a strange order which was inherited from original
debugger.lua implementation. But that was counter
intuitive, and is confusing.

Swap their meaning to be more compatible with
that we accustomed in `gdb`/`lldb`.

NO_DOC=see the later commit
NO_CHANGELOG=see the later commit