sql: throw error on creation of too many indexes
Prior to this patch, there was no error of creating too many indexes in
SQL. This led to a situation where indexes with IDs greater than
BOX_INDEX_MAX were not created, but no error was thrown. For example, in
the case of CREATE TABLE, only indexes with an ID less than
BOX_INDEX_MAX were created. Now an error is thrown when creating too
many indexes.

Closes #5526

NO_DOC=bugfix