box: move ER_UNSUPPORTED error from space_upgrade_def_decode to alter
We're planning to simplify the opts_decode function protocol: instead of
passing an error code to it and let it set ClientError, we'll rework it
so that it may only set IllegalParams error, which is supposed to be
converted to ClientError with a proper error code at the upper level.
This will let us append extra information, such as field number, to the
error.

The problem is, space_upgrade_def_decode sets ER_UNSUPPORTED error,
which is against the new protocol. Let's move the error setting to
space_upgrade_check_alter, which looks like a better place for it,
anyway. To achieve that, we make space_upgrade_def_new skip the input
and return BAD_PTR, which is used as a marker indicating that the space
upgrade options were set.

Needed for #7933

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring