vinyl: fix tuple missing in cache after reading rolled backed DELETE
The read iterator skips DELETE statements. If a skipped DELETE statement
was read from the transaction write set, the read iterator breaks the
current cache chain by clearing vy_read_iterator::last_cached, because
the deleted tuple may actually be present at a lower level so adding
a gap there until the transaction is committed would be incorrect, see
commit dd926790cc0d6 ("vinyl: fix a bug in adding of a cache chain").

The problem is vy_cache_add() assumes that if last_cached is NULL
the tuple added to the cache must be the first tuple that meets the
iteration criteria and sets the cache node boundary level accordingly.
This may lead to the previous tuple being erroneously skipped later
in case the transaction that attempted to delete it is rolled back.

Fix this issue by passing the flag indicating if the current tuple is
the first result of the iteration explicitly to vy_cache_add().

Closes #7947

NO_DOC=bug fix