backtrace: update sp before saving registers to a stack frame
backtrace_collect() is suffering from the same issue as coro_transfer(),
see commit 215630e6f6c4 ("coro: update sp before saving registers to a
stack frame").

It stores x19-x30 and d8-d15 registers to the stack, but only after that
it updates the stack pointer. If a SIGALRM signal is delivered during
the execution of backtrace_collect, the signal handler will use the
stack starting from current sp, thus corrupting the saved registers.

Fix this by updating the stack pointer at the beginning of
backtrace_collect. The constraint of the input operand `bt` is changed
from "m" to "r" to force `bt` always be passed to the inline assembly
via the register. Passing on stack will not work after this fix.

Closes #7985

NO_DOC=bugfix
NO_TEST=Hard to create a stable reproducer